Make a slight clarification.
This attempts to clarify the difference between the actual payload of an
OP_RETURN output (currently effectively 220 bytes) and the full output
sizes.
The latter includes the op_return opcode and the push opcodes.

So, simply said, this means that the new limit counts the size entire
output script, for each of the outputs using the op_return template.